+++ src/ais/mod.rs
@@ -37,6 +37,7 @@ pub(crate) mod vdm_t22;
 pub(crate) mod vdm_t23;
 pub(crate) mod vdm_t24;
 pub(crate) mod vdm_t25;
+pub(crate) mod vdm_t26;
 pub(crate) mod vdm_t27;
 
 use super::*;
@@ -55,6 +56,7 @@ pub use vdm_t21::{AidToNavigationReport, NavAidType};
 pub use vdm_t22::{ChannelManagement};
 pub use vdm_t23::{GroupAssignmentCommand};
 pub use vdm_t25::{SingleSlotBinaryMessage};
+pub use vdm_t26::{MultipleSlotBinaryMessage};
 
 // -------------------------------------------------------------------------------------------------